refactor(span): combine `Span` type and impls in 1 file (#8900)
Pure refactor. Combine all code related to `Span` into a single file.

The motivation of splitting it into 2 files originally was that `oxc_ast_tools` could not handle `impl`s in same file as type definitions. That restriction is now lifted, so there's no need for the artificial split now.
