+++ .github/.generated_ast_watch_list.yml
@@ -29,7 +29,7 @@ src:
   - 'crates/oxc_regular_expression/src/generated/derive_get_address.rs'
   - 'crates/oxc_span/src/generated/derive_estree.rs'
   - 'crates/oxc_span/src/source_type/mod.rs'
-  - 'crates/oxc_span/src/span/types.rs'
+  - 'crates/oxc_span/src/span.rs'
   - 'crates/oxc_syntax/src/generated/derive_clone_in.rs'
   - 'crates/oxc_syntax/src/generated/derive_content_eq.rs'
   - 'crates/oxc_syntax/src/generated/derive_estree.rs'
